--- SearchEngine/.gitignore
@@ -1,1 +1,0 @@
-.
