--- SearchEngine/searchEngine.log
@@ -1,36 +1,0 @@
-28-05- 2024 15:58:24 - __main__ - DEBUG - nach Masse filtern
-28-05- 2024 15:59:23 - __main__ - DEBUG - nach Masse filtern
-28-05- 2024 16:04:33 - __main__ - DEBUG - inputLoading
-28-05- 2024 16:45:43 - __main__ - DEBUG - nach Masse filtern
-28-05- 2024 17:52:24 - __main__ - DEBUG - nach Masse filtern
-28-05- 2024 17:52:39 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 08:37:01 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:38:48 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:39:28 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:41:12 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:41:24 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 08:42:40 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 08:43:11 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:43:39 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:45:05 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:51:16 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:51:36 - __main__ - DEBUG - inputLoading
-01-07- 2024 08:54:18 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 08:57:57 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 09:07:17 - __main__ - INFO - Neu geladen
-01-07- 2024 09:07:34 - __main__ - DEBUG - inputLoading
-01-07- 2024 09:09:29 - __main__ - INFO - Neu geladen
-01-07- 2024 09:09:40 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 09:10:09 - __main__ - INFO - Neu geladen
-01-07- 2024 09:10:15 - __main__ - DEBUG - inputLoading
-01-07- 2024 09:10:53 - __main__ - INFO - Neu geladen
-01-07- 2024 09:13:23 - __main__ - INFO - Neu geladen
-01-07- 2024 09:13:30 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 09:14:50 - __main__ - INFO - Neu geladen
-01-07- 2024 09:15:01 - __main__ - DEBUG - inputLoading
-01-07- 2024 09:21:36 - __main__ - INFO - Neu geladen
-01-07- 2024 09:21:44 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 09:22:03 - __main__ - DEBUG - inputLoading
-01-07- 2024 09:22:26 - __main__ - INFO - Neu geladen
-01-07- 2024 12:32:30 - __main__ - DEBUG - nach Masse filtern
-01-07- 2024 12:35:02 - __main__ - INFO - nach Masse filtern
