--- Tim_Jonas_Policija.json
@@ -1,236 +1,0 @@
-[
-    {
-        "version": "1.0",
-        "smiles": "COc1ccc2[nH]cc(CCN(C)C)c2c1",
-        "names": [
-            "5-MeO-N,N-DMT"
-        ],
-        "iupac_names": [
-            "2-(5-methoxy-1H-indol-3-yl)-N,N-dimethylethanamine"
-        ],
-        "formula": "C13H18N2O",
-        "inchi": "InChI=1S/C13H18N2O/c1-15(2)7-6-10-9-14-13-5-4-11(16-3)8-12(10)13/h4-5,8-9,14H,6-7H2,1-3H3",
-        "inchi_key": "ZSTKHSQDNIGFLM-UHFFFAOYSA-N",
-        "molecular_mass": 218.29,
-        "cas_num": null,
-        "categories": [
-            "Indolalkylamines (fe tryptamines)"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2006-01-27",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "Clc1ccc(N2CCNCC2)cc1",
-        "names": [
-            "pCPP"
-        ],
-        "iupac_names": [
-            "1-(4-chlorophenyl)piperazine"
-        ],
-        "formula": "C10H13ClN2",
-        "inchi": "InChI=1S/C10H13ClN2/c11-9-1-3-10(4-2-9)13-7-5-12-6-8-13/h1-4,12H,5-8H2",
-        "inchi_key": "UNEIHNMKASENIG-UHFFFAOYSA-N",
-        "molecular_mass": 196.67,
-        "cas_num": null,
-        "categories": [
-            "Piperazine derivates"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2009-04-30",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "Clc1cccc(N2CCNCC2)c1",
-        "names": [
-            "mCPP"
-        ],
-        "iupac_names": [
-            "1-(3-chlorophenyl)piperazine"
-        ],
-        "formula": "C10H13ClN2",
-        "inchi": "InChI=1S/C10H13ClN2/c11-9-2-1-3-10(8-9)13-6-4-12-5-7-13/h1-3,8,12H,4-7H2",
-        "inchi_key": "VHFVKMTVMIZMIK-UHFFFAOYSA-N",
-        "molecular_mass": 196.67,
-        "cas_num": null,
-        "categories": [
-            "Piperazine derivates"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2010-11-18",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "CCCCCn1cc(C(=O)Cc2ccccc2OC)c2ccccc21",
-        "names": [
-            "JWH-250"
-        ],
-        "iupac_names": [
-            "2-(2-methoxyphenyl)-1-(1-pentylindol-3-yl)ethanone"
-        ],
-        "formula": "C22H25NO2",
-        "inchi": "InChI=1S/C22H25NO2/c1-3-4-9-14-23-16-19(18-11-6-7-12-20(18)23)21(24)15-17-10-5-8-13-22(17)25-2/h5-8,10-13,16H,3-4,9,14-15H2,1-2H3",
-        "inchi_key": "FFLSJIQJQKDDCM-UHFFFAOYSA-N",
-        "molecular_mass": 335.44,
-        "cas_num": null,
-        "categories": [
-            "Cannabinoids"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2011-12-01",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "CNC(C)C(=O)c1ccc(C)cc1",
-        "names": [
-            "Mephedrone"
-        ],
-        "iupac_names": [
-            "(RS)-2-Methylamino-1-(4-methylphenyl)propan-1-one"
-        ],
-        "formula": "C11H15NO",
-        "inchi": "InChI=1S/C11H15NO/c1-8-4-6-10(7-5-8)11(13)9(2)12-3/h4-7,9,12H,1-3H3",
-        "inchi_key": "YELGFTGWJGBAQU-UHFFFAOYSA-N",
-        "molecular_mass": 177.24,
-        "cas_num": null,
-        "categories": [
-            "Cathinones"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2012-02-02",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "CNC(C)C(=O)c1cccc(C)c1",
-        "names": [
-            "3-MMC"
-        ],
-        "iupac_names": [
-            "2-(Methylamino)-1-(3-methylphenyl)-1-propanone"
-        ],
-        "formula": "C11H15NO",
-        "inchi": "InChI=1S/C11H15NO/c1-8-5-4-6-10(7-8)11(13)9(2)12-3/h4-7,9,12H,1-3H3",
-        "inchi_key": "QDNXSIYWHYGMCD-UHFFFAOYSA-N",
-        "molecular_mass": 177.24,
-        "cas_num": null,
-        "categories": [
-            "Cathinones"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2012-07-16",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "CCC(NC)C(=O)c1ccccc1",
-        "names": [
-            "Buphedrone"
-        ],
-        "iupac_names": [
-            "2-(methylamino)-1-phenylbutan-1-one"
-        ],
-        "formula": "C11H15NO",
-        "inchi": "InChI=1S/C11H15NO/c1-3-10(12-2)11(13)9-7-5-4-6-8-9/h4-8,10,12H,3H2,1-2H3",
-        "inchi_key": "DDPMGIMJSRUULN-UHFFFAOYSA-N",
-        "molecular_mass": 177.24,
-        "cas_num": null,
-        "categories": [
-            "Cathinones"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2013-05-21",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "CCC(NC)C(=O)c1ccc2c(c1)OCO2",
-        "names": [
-            "Butylone"
-        ],
-        "iupac_names": [
-            "1-(1,3-benzodioxol-5-yl)-2-(methylamino)butan-1-one"
-        ],
-        "formula": "C12H15NO3",
-        "inchi": "InChI=1S/C12H15NO3/c1-3-9(13-2)12(14)8-4-5-10-11(6-8)16-7-15-10/h4-6,9,13H,3,7H2,1-2H3",
-        "inchi_key": "CGKQZIULZRXRRJ-UHFFFAOYSA-N",
-        "molecular_mass": 221.26,
-        "cas_num": null,
-        "categories": [
-            "Cathinones"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2013-11-28",
-        "details": {}
-    },
-    {
-        "version": "1.0",
-        "smiles": "CCCC(C(=O)c1ccc2ccccc2c1)N1CCCC1",
-        "names": [
-            "Naphyrone"
-        ],
-        "iupac_names": [
-            "(RS)-1-naphthalen-2-yl-2-pyrrolidin-1-ylpentan-1-one"
-        ],
-        "formula": "C19H23NO",
-        "inchi": "InChI=1S/C19H23NO/c1-2-7-18(20-12-5-6-13-20)19(21)17-11-10-15-8-3-4-9-16(15)14-17/h3-4,8-11,14,18H,2,5-7,12-13H2,1H3",
-        "inchi_key": "",
-        "molecular_mass": 281.39,
-        "cas_num": null,
-        "categories": [
-            "Cathinones"
-        ],
-        "source": {
-            "name": "Policija",
-            "url": "https://www.policija.si/apps/nfl_response_web/seznam.php"
-        },
-        "validated": true,
-        "deleted": false,
-        "last_modified": "2013-11-28",
-        "details": {}
-    }
-]
