--- TrialAndError/paarZeilenV2.json
@@ -1,164 +1,0 @@
-[
-    {
-        "smiles": "COc1ccc2[nH]cc(CCN(C)C)c2c1",
-        "names": "5-MeO-N,N-DMT",
-        "iupac_name": "2-(5-methoxy-1H-indol-3-yl)-N,N-dimethylethanamine",
-        "formular": "C13H18N2O",
-        "molecular_mass": "218.29",
-        "Inchi": "InChI=1/C13H18N2O/c1-15(2)7-6-10-9-14-13-5-4-11(16-3)8-12(10)13/h4-5,8-9,14H,6-7H2,1-3H3",
-        "InchiKey": "ZSTKHSQDNIGFLM-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Indolalkylamines (fe tryptamines)",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2006-01-27",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": "Clc1ccc(N2CCNCC2)cc1",
-        "names": "pCPP",
-        "iupac_name": "1-(4-chlorophenyl)piperazine",
-        "formular": "C10H13ClN2",
-        "molecular_mass": "196.67",
-        "Inchi": "InChI=1/C10H13ClN2/c11-9-1-3-10(4-2-9)13-7-5-12-6-8-13/h1-4,12H,5-8H2",
-        "InchiKey": "UNEIHNMKASENIG-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Piperazine derivates",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2009-04-30",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": "Clc1cccc(N2CCNCC2)c1",
-        "names": "mCPP",
-        "iupac_name": "1-(3-chlorophenyl)piperazine",
-        "formular": "C10H13ClN2",
-        "molecular_mass": "196.67",
-        "Inchi": "InChI=1/C10H13ClN2/c11-9-2-1-3-10(8-9)13-6-4-12-5-7-13/h1-3,8,12H,4-7H2",
-        "InchiKey": "VHFVKMTVMIZMIK-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Piperazine derivates",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2010-11-18",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": "CCCCCn1cc(C(=O)Cc2ccccc2OC)c2ccccc21",
-        "names": "JWH-250",
-        "iupac_name": "2-(2-methoxyphenyl)-1-(1-pentylindol-3-yl)ethanone",
-        "formular": "C22H25NO2",
-        "molecular_mass": "335.44",
-        "Inchi": "InChI=1/C22H25NO2/c1-3-4-9-14-23-16-19(18-11-6-7-12-20(18)23)21(24)15-17-10-5-8-13-22(17)25-2/h5-8,10-13,16H,3-4,9,14-15H2,1-2H3",
-        "InchiKey": "FFLSJIQJQKDDCM-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Cannabinoids",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2011-12-01",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": "CNC(C)C(=O)c1ccc(C)cc1",
-        "names": "Mephedrone",
-        "iupac_name": "(RS)-2-Methylamino-1-(4-methylphenyl)propan-1-one",
-        "formular": "C11H15NO",
-        "molecular_mass": "177.24",
-        "Inchi": "InChI=1/C11H15NO/c1-8-4-6-10(7-5-8)11(13)9(2)12-3/h4-7,9,12H,1-3H3",
-        "InchiKey": "YELGFTGWJGBAQU-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Cathinones",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2012-02-02",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": null,
-        "names": "3-MMC",
-        "iupac_name": "2-(Methylamino)-1-(3-methylphenyl)-1-propanone",
-        "formular": "C11H15NO",
-        "molecular_mass": "177.24",
-        "Inchi": null,
-        "InchiKey": "QDNXSIYWHYGMCD-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Cathinones",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2012-07-16",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": null,
-        "names": "Buphedrone",
-        "iupac_name": "2-(methylamino)-1-phenylbutan-1-one",
-        "formular": "C11H15NO",
-        "molecular_mass": "177.24",
-        "Inchi": null,
-        "InchiKey": "DDPMGIMJSRUULN-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Cathinones",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2013-05-21",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": null,
-        "names": "Butylone",
-        "iupac_name": "1-(1,3-benzodioxol-5-yl)-2-(methylamino)butan-1-one",
-        "formular": "C12H15NO3",
-        "molecular_mass": "221.26",
-        "Inchi": null,
-        "InchiKey": "CGKQZIULZRXRRJ-UHFFFAOYSA-N",
-        "cas_num": "",
-        "category": "Cathinones",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2013-11-28",
-        "version": "1.0",
-        "details": {}
-    },
-    {
-        "smiles": null,
-        "names": "Naphyrone",
-        "iupac_name": "(RS)-1-naphthalen-2-yl-2-pyrrolidin-1-ylpentan-1-one",
-        "formular": "C19H23NO",
-        "molecular_mass": "281.39",
-        "Inchi": null,
-        "InchiKey": "",
-        "cas_num": "",
-        "category": "Cathinones",
-        "source_name": "Policija",
-        "source_url": "https://www.policija.si/apps/nfl_response_web/seznam.php",
-        "valid": null,
-        "deleted": false,
-        "last_changed_at": "2013-11-28",
-        "version": "1.0",
-        "details": {}
-    }
-]
