--- json errors
@@ -1,14 +1,0 @@
-json in iupac_name
-\u2010  -> "-" (fixed)
-\u200b  -> "" (fixed)
-
-Doesnt work:
-"\u00b2" -> "²",
-"\u2077" -> "⁷",
-"\u00b9" -> "¹",
-"\u2076" -> "⁶",
-"\u03bb" ->"λ",
-"\u03b1" -> "α",
-"\u00b1" -> "±",
-"\u2013" -> "–",
-"\u2074" -> "⁴"
